test: fix unit test build issues with new packages
Jira-Id: VPNAPPL-1775

+++ apps/macos/ProtonVPN/Extensions/Core/ProfileIcon+Extension.swift
@@ -28,7 +28,7 @@ extension ProfileIcon {
     func attributedAttachment(width: Int = 12) -> NSAttributedString {
         switch self {
         case .image(let image):
-            return image.asAttachment(style: .normal, size: .square(width))
+            return image.asAttachment(style: .normal, size: .square(CGFloat(width)))
         case .circle(let color):
             let profileCircle = ProfileCircle(frame: CGRect(x: 0, y: 0, width: width, height: width))
             profileCircle.profileColor = NSColor(rgbHex: color)
